Dump MySQL file
I added the SQL file for the back end user. The database is still being worked on, and I know there needs to be some changes done since I just realized that some of the columns I need to remove and add more. I added a unique token slot for users.